Accept container name option
When running concurrent sass compilation tasks for separate applications in a single repo, all *.scss files are copied down to the same tempDir/gulp-ruby-sass directory before being processed. This can be an issue because sass processes them en masse with a loadPath that is only relevant to one or two of the *.scss files. Unfortunately, multiple load paths won't solve the issue, because of identically named modules in each application. This commit allows a kind of name spacing of the container name to avoid the intertwining of concurrent processes.