Unique intermediate container name
This commit adds a unique intermediate container name based on loadPath to avoid the intertwining of concurrent processes with different dependencies.

When running concurrent sass compilation tasks for separate applications, all *.scss files are copied down to the same tempDir/gulp-ruby-sass directory before being processed. This can be an issue because sass processes them en masse with a loadPath that may only relevant to a subset of the *.scss files.